requirements.txt
# Core Framework
streamlit==1.31.0
python-dotenv==1.0.1

# Google Cloud AI
google-generativeai==0.3.2
google-cloud-aiplatform==1.42.0

# Data Processing
pandas==2.2.0
numpy==1.26.0
openpyxl==3.1.2

# Statistical Analysis
scipy==1.12.0
scikit-learn==1.4.0
statsmodels==0.14.1

# Visualization
plotly==5.19.0
seaborn==0.13.2
matplotlib==3.8.2

# Document Generation
quartodoc==0.7.2

# Image Processing
Pillow==10.2.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0
